requirements.txt
# Базовые зависимости
requests>=2.31.0
jsonschema>=4.18.0
python-telegram-bot>=20.3

# Поддержка тестирования
pytest>=7.4.0
pytest-mock>=3.12.0

# Поддержка анализа кода (опционально, но желательно для dev)
pylint>=3.0.0
mypy>=1.7.0

# PEP 561 совместимость с pydantic (если появится)
types-requests>=2.31.0.10
